+++ include/onmt/Tokenizer.h
@@ -38,7 +38,8 @@ namespace onmt
       CacheModel = 1024,
       SentencePieceModel = 2048,
       PreservePlaceholders = 4096,
-      SpacerNew = 8192
+      SpacerNew = 8192,
+      PreserveSegmentedTokens = 16384
     };
 
     static const std::string joiner_marker;
@@ -99,6 +100,7 @@ namespace onmt
     bool _spacer_annotate;
     bool _spacer_new;
     bool _preserve_placeholders;
+    bool _preserve_segmented_tokens;
 
     const SubwordEncoder* _subword_encoder;
     std::string _joiner;